Move InForm to InRequest and rename to InRequest
